server/requirements.txt
Flask==2.2.3
Flask_Cors==3.0.10
flask_restx==1.0.6
flask_sqlalchemy==3.0.3
numpy==1.21.6
pandas==1.3.5
PyJWT==2.6.0
sentence_transformers==2.2.2
SQLAlchemy==2.0.4